+++ controllers/providers.controller.js
@@ -1,6 +1,6 @@
 const { response } = require('express');
 const providerService = require('../services/providers.services');
-
+const {validationResult} = require("express-validator");
 
 const getProviders = async (req, res) => {
     let providers;
@@ -13,6 +13,12 @@ const getProviders = async (req, res) => {
 };
 
 const updateProviderController = async (req, res) => {
+
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+    }
+
     filter = req.query;
     update = req.body;
         try {
@@ -24,8 +30,13 @@ const updateProviderController = async (req, res) => {
 };
 
 const createProviderController = async (req, res) => {
+
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+    }
+
     const { company_name, CIF, address, url_web, isActive } = req.body;
-    if (company_name && CIF && address && url_web && isActive) {
         try {
             const response = await providerService.createProvider(company_name, CIF, address, url_web, isActive);
             res.status(201).json({
@@ -35,14 +46,17 @@ const createProviderController = async (req, res) => {
         } catch (error) {
             res.status(500).json({ mensaje: error.message });
         }
-    } else {
-        res.status(400).json({ error: "Faltan campos de provider" });
-    }
 };
 
 // deleteProvider
 // DELETE http://localhost:3000/api/providers?company_name=name
 const deleteProviderController = async (req, res) => {
+
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+    }
+
     let providers;
     try {
         providers = await providerService.deleteProvider(req.query.company_name);
@@ -52,20 +66,20 @@ const deleteProviderController = async (req, res) => {
     }
 };
 
-const toggleProviderController = async (req, res) => {
-    const company_name = req.query.company_name;
-    if (company_name) {
-        try {
-            const response = await providerService.toggleProvider(company_name);
-            res.status(200).json(response);
-        } catch (error) {
-            res.status(500).json(response);
-        }
+// const toggleProviderController = async (req, res) => {
+//     const company_name = req.query.company_name;
+//     if (company_name) {
+//         try {
+//             const response = await providerService.toggleProvider(company_name);
+//             res.status(200).json(response);
+//         } catch (error) {
+//             res.status(500).json(response);
+//         }
 
-    } else {
-        res.status(400).json({ error: "Company name is required" });
-    }
-};
+//     } else {
+//         res.status(400).json({ error: "Company name is required" });
+//     }
+// };
 
 
 module.exports = {
@@ -73,5 +87,5 @@ module.exports = {
     createProviderController,
     updateProviderController,
     deleteProviderController,
-    toggleProviderController
+    // toggleProviderController
 };
